Simplified implementation of listener and get/set methods.
